move nut19 to nuts directory
